+++ auth/auth.go
@@ -16,7 +16,6 @@
 package auth
 
 import (
-	"context"
 	"crypto/rsa"
 	"crypto/x509"
 	"encoding/json"
@@ -24,6 +23,8 @@ import (
 	"errors"
 	"fmt"
 	"strings"
+
+	"golang.org/x/net/context"
 
 	"firebase.google.com/go/internal"
 	"google.golang.org/api/identitytoolkit/v3"
@@ -62,7 +63,6 @@ type Token struct {
 // Client facilitates generating custom JWT tokens for Firebase clients, and verifying ID tokens issued
 // by Firebase backend services.
 type Client struct {
-	hc        *internal.HTTPClient
 	is        *identitytoolkit.Service
 	ks        keySource
 	projectID string
@@ -123,7 +123,6 @@ func NewClient(ctx context.Context, c *internal.AuthConfig) (*Client, error) {
 	}
 
 	return &Client{
-		hc:        &internal.HTTPClient{Client: hc},
 		is:        is,
 		ks:        newHTTPKeySource(googleCertURL, hc),
 		projectID: c.ProjectID,
@@ -266,7 +265,7 @@ func (c *Client) VerifyIDTokenAndCheckRevoked(ctx context.Context, idToken strin
 	}
 
 	if p.IssuedAt*1000 < user.TokensValidAfterMillis {
-		return nil, fmt.Errorf("ID token has been revoked")
+		return nil, internal.Error(idTokenRevoked, "ID token has been revoked")
 	}
 	return p, nil
 }