+++ auth/auth_std.go
@@ -14,9 +14,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package auth
+package auth // import "firebase.google.com/go/auth"
 
-import "context"
+import "golang.org/x/net/context"
 
 func newSigner(ctx context.Context) (signer, error) {
 	return serviceAcctSigner{}, nil