+++ firebase.go
@@ -15,14 +15,15 @@
 // Package firebase is the entry point to the Firebase Admin SDK. It provides functionality for initializing App
 // instances, which serve as the central entities that provide access to various other Firebase services exposed
 // from the SDK.
-package firebase
+package firebase // import "firebase.google.com/go"
 
 import (
-	"context"
 	"encoding/json"
 	"errors"
 	"io/ioutil"
 	"os"
+
+	"golang.org/x/net/context"
 
 	"cloud.google.com/go/firestore"
 
@@ -41,7 +42,7 @@ import (
 var defaultAuthOverrides = make(map[string]interface{})
 
 // Version of the Firebase Go Admin SDK.
-const Version = "2.6.0"
+const Version = "2.6.1"
 
 // firebaseEnvName is the name of the environment variable with the Config.
 const firebaseEnvName = "FIREBASE_CONFIG"