+++ integration/messaging/messaging_test.go
@@ -15,12 +15,13 @@
 package messaging
 
 import (
-	"context"
 	"flag"
 	"log"
 	"os"
 	"regexp"
 	"testing"
+
+	"golang.org/x/net/context"
 
 	"firebase.google.com/go/integration/internal"
 	"firebase.google.com/go/messaging"
@@ -88,7 +89,7 @@ func TestSend(t *testing.T) {
 	}
 	name, err := client.SendDryRun(context.Background(), msg)
 	if err != nil {
-		log.Fatalln(err)
+		t.Fatal(err)
 	}
 	const pattern = "^projects/.*/messages/.*$"
 	if !regexp.MustCompile(pattern).MatchString(name) {