+++ integration/storage/storage_test.go
@@ -15,13 +15,14 @@
 package storage
 
 import (
-	"context"
 	"flag"
 	"fmt"
 	"io/ioutil"
 	"log"
 	"os"
 	"testing"
+
+	"golang.org/x/net/context"
 
 	"firebase.google.com/go"